# sql: issue GetRequests from SQL when possible