--- a/pkg/sql/opt/exec/execbuilder/testdata/autocommit_nonmetamorphic
+++ b/pkg/sql/opt/exec/execbuilder/testdata/autocommit_nonmetamorphic
@@ -668,7 +668,7 @@ WHERE message       LIKE '%r$rangeid: sending batch%'
 ----
 dist sender send  r37: sending batch 1 DelRng to (n1,s1):1
 dist sender send  r37: sending batch 2 CPut to (n1,s1):1
-dist sender send  r37: sending batch 2 Scan to (n1,s1):1
+dist sender send  r37: sending batch 2 Get to (n1,s1):1
 dist sender send  r37: sending batch 1 EndTxn to (n1,s1):1
 
 query B
@@ -717,7 +717,7 @@ WHERE message       LIKE '%r$rangeid: sending batch%'
   AND operation NOT LIKE '%async%'
 ----
 dist sender send  r37: sending batch 1 DelRng to (n1,s1):1
-dist sender send  r37: sending batch 1 Scan to (n1,s1):1
+dist sender send  r37: sending batch 1 Get to (n1,s1):1
 dist sender send  r37: sending batch 1 Del to (n1,s1):1
 dist sender send  r37: sending batch 1 Scan to (n1,s1):1
 dist sender send  r37: sending batch 1 EndTxn to (n1,s1):1