… pkg/ccl/logictestccl/testdata/logic_test/multi_region
… pkg/kv/kvserver/batcheval/cmd_get_test.go
… pkg/kv/kvserver/replica_evaluate_test.go
… pkg/sql/as_of_test.go
… pkg/sql/logictest/testdata/logic_test/explain_analyze_plans
… pkg/sql/logictest/testdata/logic_test/inverted_filter_json_array_dist
… pkg/sql/logictest/testdata/logic_test/merge_join_dist_vec
… pkg/sql/opt/exec/execbuilder/testdata/orderby
… pkg/sql/opt/exec/execbuilder/testdata/partial_index_nonmetamorphic
… pkg/sql/opt/exec/execbuilder/testdata/secondary_index_column_families_nonmetamorphic
… pkg/sql/opt/exec/execbuilder/testdata/select
… pkg/sql/opt/exec/execbuilder/testdata/select_for_update
… pkg/sql/opt/exec/execbuilder/testdata/unique
… pkg/sql/opt/exec/execbuilder/testdata/upsert
… pkg/sql/opt/exec/execbuilder/testdata/upsert_nonmetamorphic
… pkg/sql/opt/exec/execbuilder/testdata/virtual_columns
… pkg/sql/row/fk_spans.go
… pkg/sql/row/kv_batch_fetcher.go
… pkg/sql/rowenc/index_encoding.go
… pkg/sql/span/span_builder.go
… pkg/sql/span/span_builder_test.go
… pkg/sql/span_builder_test.go
… pkg/sql/testdata/explain_tree
… pkg/sql/txn_restart_test.go
… pkg/sql/upsert_test.go