@jordanlewis 2021-03-06
So far, this looks to make almost no difference, but I haven't spent time doing a careful benchmark, and I don't know the best/worst cases to test against.

I ran KV95 on a single gceworker node for 60 seconds on empty data directories before and after the patch.

After the patch:

```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   60.0s        0        2039148        33985.5      0.9      0.7      1.8      5.0     35.7  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   60.0s        0         106947         1782.4     10.2     10.0     17.8     21.0     41.9  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
   60.0s        0        2146095        35767.9      1.3      0.7      6.3     13.6     41.9
```


Before the patch:

```
_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   60.0s        0        2024099        33734.7      0.9      0.7      2.0      5.0     35.7  read

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__total
   60.0s        0         106675         1777.9      9.9      9.4     16.8     21.0     39.8  write

_elapsed___errors_____ops(total)___ops/sec(cum)__avg(ms)__p50(ms)__p95(ms)__p99(ms)_pMax(ms)__result
   60.0s        0        2130774        35512.6      1.3      0.7      6.0     13.1     39.8
```