@jordanlewis 2021-03-27
@tbg or @nvanbenschoten, PTAL at the first commit, which teaches `cmd_get` to leave early if it runs out of bytes or key quota. It's entirely possible that this is undesired behavior, I have no idea :) The reason that I added it is that, now that SQL can emit mixed scan and get requests, it will receive out of order results unless something like this change is added.

Consider a 4 column family table and a query that skips an inner column family.

```
CREATE TABLE a (a INT PRIMARY KEY, b INT, c INT, d INT, FAMILY(a), FAMILY(b), FAMILY(c), FAMILY(d)
SELECT a, b, d FROM a WHERE a = 10
```

This PR handles this request by creating a ScanRequest for `a,b`, and a GetRequest for `d`. If the ScanRequest runs out of quota during `a,b`, and returns a `ResumeSpan` for just `b`, but the `GetRequest` ignores the key/bytes limit and returns a result unconditionally, then SQL ends up seeing the keys in order `a`, `d`, `b`, which is invalid and causes various things to break.

I suppose we could fix this in another way, like teaching the distsender or fetcher to be smarter somehow, but that seems inelegant - to me it feels natural that `GetRequest` would obey the bytes/keys limits the same way that `ScanRequest` does.